requirements.txt
ansible-core
redbaron
ruamel.yaml
antsibull-docs-parser >= 1.0.0, < 2.0.0
